+++ lib/merge_params/helpers.rb
@@ -1,25 +1,30 @@
 require 'uri'
+require 'facets/hash/deep_merge'
+require 'facets/hash/recurse'
+require 'facets/hash/recursively_comparing'
+require 'facets/enumerable/graph'
+require 'facets/hash/graph'
 
 module MergeParams::Helpers
   extend ActiveSupport::Concern
 
   # request.parameters but with symbolized keys.
   def request_params
-    request.parameters.symbolize_keys
+    request.parameters.deep_symbolize_keys
   end
 
   # request.parameters (which also includes POST params) but with only those keys that would
   # normally be passed in a query string (without :controller, :action, :format) and with symbolized
   # keys.
   def query_params_from_request_params
-    request.parameters.symbolize_keys.
-      except(*request.path_parameters.symbolize_keys.keys)
+    request.parameters.deep_symbolize_keys.
+      except(*request.path_parameters.deep_symbolize_keys.keys)
   end
 
   # Returns a hash of params from the query string (https://en.wikipedia.org/wiki/Query_string),
   # with symbolized keys.
   def query_params
-    request.query_parameters.symbolize_keys
+    request.query_parameters.deep_symbolize_keys
   end
 
   # Params that can safely be passed to url_for to build a route. (Used by merge_url_for.)
@@ -40,8 +45,9 @@ module MergeParams::Helpers
   # (And we don't even need to pass the path_parameters on to url_for because url_for already
   # includes those (from :_recall)
   #
-  def params_for_url_for
-    params.to_unsafe_h.symbolize_keys.except(
+  def params_for_url_for(params = params())
+    params = params.to_unsafe_h if params.respond_to?(:to_unsafe_h)
+    params.deep_symbolize_keys.except(
       *ActionDispatch::Routing::RouteSet::RESERVED_OPTIONS,
       :controller,
       :action,
@@ -51,7 +57,8 @@ module MergeParams::Helpers
 
   # Safely merges the given params with the params from the current request
   def merge_params(new_params = {})
-    params_for_url_for.merge(new_params)
+    params_for_url_for.
+      merge(new_params.deep_symbolize_keys)
   end
 
   # Easily extract just certain param keys.
@@ -83,31 +90,31 @@ module MergeParams::Helpers
   def merge_url_for(new_params = {})
     url = url_for(merge_params(new_params))
 
-    # Now pass along in the *query string* any params that we couldn't pass to url_for because they
-    # were reserved options.
-    query_params_already_added = parse_nested_query(URI(url).query || '')
-    # Some params from new_params (like company_id) that we pass in may be recognized by a route and
-    # therefore no longer be query params. We use recognize_path to find those params that ended up
-    # as route params instead of query_params but are nonetheless aready added to the url.
-    params_already_added = Rails.application.routes.recognize_path(url).merge(query_params_already_added)
-    keys_already_added = params_already_added.keys
-    # Allow keys that are currently in query_params to be deleted by setting their value to nil in
-    # new_params.
-    keys_to_delete = new_params.select {|k,v| v.nil?}.keys
-    query_params_to_add = query_params.except(*keys_already_added + keys_to_delete)
+#    # Now pass along in the *query string* any params that we couldn't pass to url_for because they
+#    # were reserved options.
+#    query_params_already_added = parse_nested_query(URI(url).query || '')
+#    # Some params from new_params (like company_id) that we pass in may be recognized by a route and
+#    # therefore no longer be query params. We use recognize_path to find those params that ended up
+#    # as route params instead of query_params but are nonetheless already added to the url.
+#    params_already_added = Rails.application.routes.recognize_path(url).merge(query_params_already_added)
+    params_already_added = params_from_url(url)
+    query_params_to_add = new_params.recursively_comparing(params_already_added).graph { |k,v, other|
+      if v.is_a?(Hash) || v.nil? || other.nil?
+        [k, v]
+      end
+    }
     add_params(query_params_to_add, url)
   end
 
   # Adds params to the query string
   # (Unlike url_for_merge, which tries to generate a route from the params.)
-  # TODO: Should URL be first like https://libraries.io/github/jordanmaguire/uri_query_merger ?
-  #   UriQueryMerger.new("http://www.google.com?other=1", {jordan: "rules"}).merge
-  # Can we make it work that way when a URL is supplied buth otherwise let the params be the first
-  # and only argument (to optimize for that more common use case)?
-  def add_params(params = {}, url = request.fullpath)
+  def add_params(new_params = {}, url = request.fullpath)
     uri = URI(url)
-    params    = parse_nested_query(uri.query || '').merge(params)
-    uri.query = Rack::Utils.build_nested_query(params) if params.present?
+    # Allow keys that are currently in query_params to be deleted by setting their value to nil in
+    # new_params (including in nested hashes).
+    merged_params = parse_nested_query(uri.query || '').
+      deep_merge(new_params).recurse(&:compact)
+    uri.query = Rack::Utils.build_nested_query(merged_params).presence
     uri.to_s
   end
 
@@ -121,9 +128,16 @@ module MergeParams::Helpers
     ) if respond_to?(:helper_method)
   end
 
-private
+  # Parsing helpers
+  def params_from_url(url)
+    query_params = parse_nested_query(URI(url).query || '')
+    route_params = Rails.application.routes.recognize_path(url.to_s)
+    params_for_url_for(
+      route_params.merge(query_params)
+    )
+  end
 
   def parse_nested_query(query)
-    Rack::Utils.parse_nested_query(query || '').symbolize_keys
+    Rack::Utils.parse_nested_query(query || '').deep_symbolize_keys
   end
 end